requirements.txt
pandas
requests
pyyaml
pyarrow
openpyxl
matplotlib
duckdb
